split integration tests to reduce test duration
