+++ tests/integration/ha/upgrades/literals.py
@@ -1,0 +1,11 @@
+#!/usr/bin/env python3
+# Copyright 2025 Canonical Ltd.
+# See LICENSE file for licensing details.
+
+"""Collection of literals for upgrades tests."""
+
+NUM_UNITS = 3
+CHARM_CHANNEL = "3.6/edge"
+CHARM_REVISIONS_TO_DEPLOY = {"x86_64": 89, "aarch64": 88}
+WORKLOAD_VERSION = {"previous": "3.6.1", "target": "3.6.2"}
+CERTIFICATE_EXPIRY_TIME = 250
