+++ tests/spread/test_upgrade_cluster_size_related_tests.py/task.yaml
@@ -1,6 +1,6 @@
 summary: test_upgrade_edge_cases.py
 environment:
-  TEST_MODULE: ha/upgrades/test_upgrade_edge_cases.py
+  TEST_MODULE: ha/upgrades/test_upgrade_cluster_size_related_tests.py
 systems:
   - self-hosted-linux-amd64-noble-large
   - self-hosted-linux-arm64-noble-large
